requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# rapid_react_scouting/main.py: 1
uvicorn == 0.16.0
fastapi == 0.71.0
deta
python-jose
passlib
pydantic
python-multipart
argon2-cffi